requirements.txt
aiohttp==3.9.5
aiosignal==1.2.0
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
attrs==23.1.0
Bottleneck==1.3.7
Brotli==1.0.9
certifi==2024.6.2
charset-normalizer==2.0.4
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.1
debugpy==1.6.7
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.53.0
frozenlist==1.4.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
huggingface_hub==0.23.3
idna==3.7
importlib_metadata==7.1.0
inquirerpy==0.3.4
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.5.0
kiwisolver==1.4.5
lightning==2.2.5
lightning-utilities==0.11.2
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nest_asyncio==1.6.0
networkx==3.2.1
numexpr==2.8.7
numpy==1.26.4
opencv-python==4.10.0
opencv-python-headless==4.10.0
packaging==23.2
pandas==2.2.1
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
pfzy==0.3.4
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.10.3
requests==2.32.2
safetensors==0.4.0
sentry-sdk==2.5.0
setproctitle==1.2.2
setuptools==69.5.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
sympy==1.12
tiktoken==0.7.0
tokenizers==0.15.1
torch==2.2.2
torchmetrics==1.4.0.post0
torchvision==0.17.2
tornado==6.3.3
tqdm==4.66.4
traitlets==5.14.3
transformers==4.37.1
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.1
wandb==0.16.5
wcwidth==0.2.13
wheel==0.43.0
xxhash==2.0.2
yarl==1.9.3
zipp==3.18.1
